Тесты API вынесены в пакет tests/api, экспорт-тесты разделены по слою
