+++ tests/api/test_periods.py
